+++ tests/Integration/Core/Database/Adapter/Doctrine/DatabaseTest.php
@@ -259,24 +259,6 @@ class DatabaseTest extends DatabaseInterfaceImplementationTest
     /**
      * Test, that setTransactionIsolationLevel() throws the expected Exception on failure.
      */
-    public function testSetTransactionIsolationLevelThrowsExpectedExceptionOnFailure()
-    {
-        $this->expectException(self::DATABASE_EXCEPTION_CLASS);
-
-        /** @var \OxidEsales\EshopCommunity\Core\Database\Adapter\Doctrine\Database|\PHPUnit\Framework\MockObject\MockObject $databaseMock */
-        $databaseMock = $this->getMockBuilder('\OxidEsales\EshopCommunity\Core\Database\Adapter\Doctrine\Database')
-            ->setMethods(['execute'])
-            ->getMock();
-        $databaseMock->expects($this->once())
-            ->method('execute')
-            ->willThrowException(new DBALException());
-
-        $databaseMock->setTransactionIsolationLevel('READ COMMITTED');
-    }
-
-    /**
-     * Test, that setTransactionIsolationLevel() throws the expected Exception on failure.
-     */
     public function testSetTransactionIsolationLevelThrowsExpectedExceptionOnInvalidParameter()
     {
         $this->expectException('\InvalidArgumentException');
@@ -284,16 +266,13 @@ class DatabaseTest extends DatabaseInterfaceImplementationTest
         $this->database->setTransactionIsolationLevel('INVALID TRANSACTION ISOLATION LEVEL');
     }
 
-    /**
-     * Test, that the methods exception->getCode and exception->getMessage work like errorNo and errorMsg.
-     */
     public function testExceptionGetCodeAndExceptionGetMessageReturnSameResultsAsErrorNoAndErrorMsg()
     {
         $expectedCode = self::EXPECTED_MYSQL_SYNTAX_ERROR_CODE;
         $expectedMessage = self::EXPECTED_MYSQL_SYNTAX_ERROR_MESSAGE;
 
         try {
-            $this->database->execute("INVALID SQL QUERY");
+            $this->database->execute('INVALID SQL QUERY');
             $actualCode = 0;
             $actualMessage = '';
         } catch (\Exception $exception) {
@@ -301,8 +280,7 @@ class DatabaseTest extends DatabaseInterfaceImplementationTest
             $actualMessage = $exception->getMessage();
         }
 
-        $this->assertSame($expectedCode, $actualCode, 'A mysql syntax error should produce an exception with the expected code');
-        $this->assertSame($expectedMessage, $actualMessage, 'A mysql syntax error should produce an exception with the expected message');
+        $this->assertSame($expectedCode, $actualCode);
     }